+++ app/src/main/java/com/example/bininfotracker/domain/model/CardInfo.kt
@@ -3,8 +3,8 @@ package com.example.bininfotracker.domain.model
 data class CardInfo(
     val scheme: String?,
     val countryName: String?,
-    val latitude: Long?,
-    val longitude: Long?,
+    val latitude: Int?,
+    val longitude: Int?,
     val bankName: String?,
     val bankUrl: String?,
     val bankPhone: String?,
